requirements.txt
tensorflow
matplotlib
numpy
pillow
scikit-learn
seaborn
kagglehub
pandas
ipykernel